refactor: Moves ClientConnectionServer and WebsocketServer to data-access layer.
